requirements.txt
tdl-client-python==0.30.1
coverage==7.8.0
pytest==8.2.2
pytest-cov==6.0.0
